Feat: blockchain.json and others

+++ src/mining/src/transaction.rs
@@ -1,6 +1,9 @@
-#[derive(Debug)]
+// transaction.rs
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Transaction {
     pub sender: String,
     pub recipient: String,
-    pub amount: u64,
+    pub amount: u32,
 }
